@@ -1,9 +1,7 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:input_widgets_demo/inputs_demo/bloc/real_estate_search_form_cubit.dart';
 
 class CheckboxRow extends StatelessWidget {
-  const CheckboxRow({Key? key}) : super(key: key);
+  const CheckboxRow({super.key});
 
   @override
   Widget build(BuildContext context) {
