@@ -1,11 +1,7 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:input_widgets_demo/inputs_demo/bloc/real_estate_search_form_cubit.dart';
-
-import '../../../util.dart';
 
 class RadioGroup extends StatelessWidget {
-  const RadioGroup({Key? key}) : super(key: key);
+  const RadioGroup({super.key});
 
   @override
   Widget build(BuildContext context) {
